requirements.txt
pyttsx3
SpeechRecognition
Pillow
PyAudio